modified 1-print_binary.c

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -8,7 +8,7 @@
  */
 void print_binary(unsigned long int n)
 {
-	int ref = n, len = 0, num = n;
+	unsigned long int ref = n, len = 0, num = n;
 
 	if (n == 0)
 		_putchar('0');
